+++ src/main/java/net/frozenblock/wilderwild/block/impl/BubbleDirection.java
@@ -16,22 +16,24 @@
  * along with this program; if not, see <https://www.gnu.org/licenses/>.
  */
 
-package net.frozenblock.wilderwild.block.property;
+package net.frozenblock.wilderwild.block.impl;
 
+import java.util.Optional;
+import net.minecraft.core.Direction;
 import net.minecraft.util.StringRepresentable;
 import org.jetbrains.annotations.NotNull;
 
-public enum FlowerColor implements StringRepresentable {
-	NONE("none"),
-	BLUE("blue"),
-	WHITE("white"),
-	PINK("pink"),
-	PURPLE("purple");
+public enum BubbleDirection implements StringRepresentable {
+	NONE("none", Optional.empty()),
+	UP("up", Optional.of(Direction.UP)),
+	DOWN("down", Optional.of(Direction.DOWN));
 
+	public final Optional<Direction> direction;
 	private final String name;
 
-	FlowerColor(String name) {
+	BubbleDirection(String name, Optional<Direction> direction) {
 		this.name = name;
+		this.direction = direction;
 	}
 
 	@Override