+++ src/main/java/net/frozenblock/wilderwild/mixin/sculk/SculkSpreaderChargeCursorMixin.java
@@ -25,7 +25,7 @@ import com.llamalad7.mixinextras.sugar.ref.LocalBooleanRef;
 import java.util.Iterator;
 import java.util.List;
 import net.frozenblock.lib.sculk.api.BooleanPropertySculkBehavior;
-import net.frozenblock.wilderwild.block.sculk_behavior.SlabWallStairSculkBehavior;
+import net.frozenblock.wilderwild.block.impl.SlabWallStairSculkBehavior;
 import net.frozenblock.wilderwild.registry.RegisterBlocks;
 import net.frozenblock.wilderwild.registry.RegisterProperties;
 import net.frozenblock.wilderwild.tag.WilderBlockTags;