+++ src/main/java/net/frozenblock/wilderwild/mixin/snowlogging/BoneMealItemMixin.java
@@ -1,0 +1,52 @@
+/*
+ * Copyright 2024 FrozenBlock
+ * This file is part of Wilder Wild.
+ *
+ * This program is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU Lesser General Public
+ * License as published by the Free Software Foundation; either
+ * version 3 of the License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with this program; if not, see <https://www.gnu.org/licenses/>.
+ */
+
+package net.frozenblock.wilderwild.mixin.snowlogging;
+
+import net.frozenblock.wilderwild.block.impl.SnowloggingUtils;
+import net.minecraft.core.BlockPos;
+import net.minecraft.world.InteractionResult;
+import net.minecraft.world.item.BoneMealItem;
+import net.minecraft.world.item.context.UseOnContext;
+import net.minecraft.world.level.Level;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
+import org.spongepowered.asm.mixin.injection.callback.LocalCapture;
+
+@Mixin(BoneMealItem.class)
+public class BoneMealItemMixin {
+
+	@Inject(method = "useOn",
+		at = @At(
+			value = "INVOKE",
+			target = "Lnet/minecraft/world/item/BoneMealItem;growCrop(Lnet/minecraft/world/item/ItemStack;Lnet/minecraft/world/level/Level;Lnet/minecraft/core/BlockPos;)Z",
+			ordinal = 0,
+			shift = At.Shift.BEFORE
+		),
+		cancellable = true,
+		locals = LocalCapture.CAPTURE_FAILHARD
+	)
+	public void wilderWild$useOn(UseOnContext context, CallbackInfoReturnable<InteractionResult> info, Level level, BlockPos blockPos, BlockPos blockPos2) {
+		if (SnowloggingUtils.isSnowlogged(level.getBlockState(blockPos))) {
+			info.setReturnValue(InteractionResult.PASS);
+		}
+	}
+
+}