+++ src/main/java/net/frozenblock/wilderwild/mixin/snowlogging/SnowyDirtBlockMixin.java
@@ -20,7 +20,7 @@ package net.frozenblock.wilderwild.mixin.snowlogging;
 
 import com.llamalad7.mixinextras.injector.wrapoperation.Operation;
 import com.llamalad7.mixinextras.injector.wrapoperation.WrapOperation;
-import net.frozenblock.wilderwild.registry.RegisterProperties;
+import net.frozenblock.wilderwild.block.impl.SnowloggingUtils;
 import net.minecraft.tags.TagKey;
 import net.minecraft.world.level.block.SnowyDirtBlock;
 import net.minecraft.world.level.block.state.BlockState;
@@ -32,7 +32,7 @@ public class SnowyDirtBlockMixin {
 
 	@WrapOperation(method = "isSnowySetting", at = @At(value = "INVOKE", target = "Lnet/minecraft/world/level/block/state/BlockState;is(Lnet/minecraft/tags/TagKey;)Z"))
 	private static boolean wilderWild$isSnowySetting(BlockState instance, TagKey tagKey, Operation<Boolean> original) {
-		return original.call(instance, tagKey) || (RegisterProperties.isSnowlogged(instance));
+		return original.call(instance, tagKey) || (SnowloggingUtils.isSnowlogged(instance));
 	}
 
 }