+++ src/main/java/net/frozenblock/wilderwild/registry/RegisterProperties.java
@@ -18,17 +18,11 @@
 
 package net.frozenblock.wilderwild.registry;
 
-import net.frozenblock.wilderwild.block.property.BubbleDirection;
-import net.frozenblock.wilderwild.block.property.FlowerColor;
-import net.minecraft.world.level.block.Blocks;
-import net.minecraft.world.level.block.BushBlock;
-import net.minecraft.world.level.block.state.BlockState;
-import net.minecraft.world.level.block.state.properties.BlockStateProperties;
+import net.frozenblock.wilderwild.block.impl.BubbleDirection;
+import net.frozenblock.wilderwild.block.impl.FlowerColor;
 import net.minecraft.world.level.block.state.properties.BooleanProperty;
 import net.minecraft.world.level.block.state.properties.EnumProperty;
 import net.minecraft.world.level.block.state.properties.IntegerProperty;
-import org.jetbrains.annotations.NotNull;
-import org.spongepowered.asm.mixin.Unique;
 
 public final class RegisterProperties {
 
@@ -78,27 +72,6 @@ public final class RegisterProperties {
 	}
 
 	public static void init() {
-	}
-
-	public static boolean canBeSnowlogged(@NotNull BlockState state) {
-		return state.hasProperty(SNOW_LAYERS);
-	}
-
-	public static int getSnowLayers(@NotNull BlockState state) {
-		return state.getValue(SNOW_LAYERS);
-	}
-
-	public static boolean isSnowlogged(BlockState state) {
-		return canBeSnowlogged(state) && getSnowLayers(state) > 0;
-	}
-
-	@NotNull
-	public static BlockState getSnowEquivalent(BlockState state) {
-		return Blocks.SNOW.defaultBlockState().setValue(BlockStateProperties.LAYERS, Math.max(1, getSnowLayers(state)));
-	}
-
-	public static boolean isBreakableWithSnow(@NotNull BlockState state) {
-		return state.getBlock() instanceof BushBlock;
 	}
 
 }