+++ CHANGELOG.md
@@ -11,3 +11,4 @@ Bug Fixes & Changes
   - The `wilderwild:small_sponge_grows_on` tag now includes the `wilderwild:mesoglea` tag instead of only the two Pearlescent Mesoglea types.
   - Restricted the placement of multiple features, so they will no longer generate in unwanted places, like on top of structures.
     - Added the `wilderwild:fallen_tree_placeable` tag to control where Fallen Trees can generate.
+  - Refactored the `BubbleDirection,` `FlowerColor,` and `SlabWillStairSculkBehavior` classes to the `block.impl` package.